requirements.txt
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
certifi==2023.11.17
distlib==0.3.8
distro==1.9.0
Django==5.1
filelock==3.13.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.8
jiter==0.5.0
openai==1.42.0
platformdirs==4.1.0
pydantic==2.8.2
pydantic_core==2.20.1
python-decouple==3.8
setuptools==72.1.0
sniffio==1.3.1
sqlparse==0.5.1
tqdm==4.66.5
typing_extensions==4.12.2
virtualenv==20.25.0
wheel==0.43.0
